+++ tests/bootstrap.php
@@ -39,4 +39,8 @@ function _manually_load_plugin() {
 }
 tests_add_filter( 'muplugins_loaded', '_manually_load_plugin' );
 
+if ( ! defined( 'WP_TESTS_CONFIG_FILE_PATH' ) && file_exists( __DIR__ . '/wp-tests-config.php' ) ) {
+	define( 'WP_TESTS_CONFIG_FILE_PATH', __DIR__ . '/wp-tests-config.php' );
+}
+
 require $test_root . '/includes/bootstrap.php';